+++ Changes.md
@@ -4,6 +4,10 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
+## HEAD
+
+* Fix timing of fatal protocol errors bubbling to user #879
+
 ## v2.4.1 (2014-07-17)
 
 * Fix `pool.query` not invoking callback on connection error #872